requirements.txt
Django==1.6.5
Markdown==2.4.1
Pillow==2.4.0
PyPDF2==1.21
dj-database-url==0.3.0
dj-static==0.0.6
django-filter==0.7
django-toolbelt==0.0.1
djangorestframework==2.3.14
gunicorn==19.1.0
psycopg2==2.5.3
requests==2.3.0
static3==0.5.1
wsgiref==0.1.2
